Make prettier actually use the config files (#276)
Previously, there were two config files for prettier.
`.prettierrc.json` was empty, whereas `.prettierrc.toml` contained
the actual configuration.  However, Prettier was reading the empty
`.prettierrc.json` by default, meaning that all configuration in
`.prettierrc.toml` was ignored.

This PR fixes this, and also adds a tab-stop of 4 to the configuration.
This way, the formatting is not affected by editor preferences.